2.0.0 split simple4h

See merge request web/myopia-pac/myopia-pac-backend!37

+++ myopia-business/business-api/screening-app-api/src/main/java/com/wupol/myopia/business/api/screening/app/service/ScreeningAppService.java
@@ -327,7 +327,7 @@ public class ScreeningAppService {
      * @return
      */
     public List<School> getSchoolByScreeningOrgId(Integer screeningOrgId) {
-        List<Long> schoolIds = screeningPlanService.getScreeningSchoolIdByScreeningOrgId(screeningOrgId);
+        List<Integer> schoolIds = screeningPlanService.getScreeningSchoolIdByScreeningOrgId(screeningOrgId);
         return schoolService.getSchoolByIds(schoolIds);
     }
